requirements.txt
# Core dependencies
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Audio/Video processing
ffmpeg-python>=0.2.0

# Configuration and templating
PyYAML>=6.0
Jinja2>=3.1.0

# CLI and utilities
click>=8.0.0
rich>=13.0.0
tqdm>=4.65.0

# File handling
pathlib2>=2.3.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0

# Optional: GPU acceleration
# torch-audio>=2.0.0+cu118  # Uncomment for CUDA support

# Development dependencies
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
